requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
astroid==2.5
awesome-django-timezones==0.3.0
Brotli==1.0.9
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
cryptography==3.4.8
cssselect2==0.4.1
decorator==5.1.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.10
django-import-export==2.5.0
django-multiselectfield==0.1.12
django-tinymce==3.3.0
et-xmlfile==1.1.0
fonttools==4.28.1
future==0.18.2
html5lib==1.1
idna==3.2
importlib-metadata==4.8.1
ipapi==1.0.4
isort==5.9.3
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe
mkl-fft
mkl-random
mkl-service
munkres
mysqlclient==2.0.3
numpy==1.20.1
odfpy
olefile
openpyxl
pandas==1.2.3
Pillow==8.1.0
pip==21.1.3
pycparser
pydyf
pylint
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pyphen==0.11.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
ratelim==0.1.6
reportlab==3.6.2
requests==2.26.0
setuptools==52.0.0.post20211006
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
toml==0.10.1
urllib3==1.26.6
weasyprint==52.5
webencodings==0.5.1
wheel==0.36.2
whitenoise==5.2.0
wincertstore==0.2
wrapt
xlrd
xlwt
zipp
zopfli
